% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dmatrices.R
\name{generate_dmatrices}
\alias{generate_dmatrices}
\title{generate_dmatrices}
\usage{
generate_dmatrices(files, infofile)
}
\arguments{
\item{files}{...}

\item{infofile}{...}
}
\value{
...
}
\description{
generate_dmatrices
}
