% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_glist.R
\name{generate_glist}
\alias{generate_glist}
\title{generate_glist}
\usage{
generate_glist(files, infofile)
}
\arguments{
\item{files}{...}

\item{infofile}{...}
}
\value{
...
}
\description{
generate_glist
}
