% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parser.R
\name{parse_dmatrices}
\alias{parse_dmatrices}
\title{Parsing inputs to data matricies}
\usage{
parse_dmatrices(files, infotable)
}
\arguments{
\item{files}{vector of character strings representing paths to all CRISPR
single-screen results.}

\item{infotable}{data.frame containing information about the individual
CRISPR screen results.}
}
\value{
list of two matricies
}
\description{
Parsing input files to the matrix format used by the empirical Brown's
method.
}
