% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parser.R
\name{parse_glist}
\alias{parse_glist}
\title{Parsing inputs to genelists}
\usage{
parse_glist(files, infotable)
}
\arguments{
\item{files}{vector of character strings representing paths to all CRISPR
single-screen results.}

\item{infotable}{data.frame containing information about the individual
CRISPR screen results.}
}
\value{
list of genelists
}
\description{
Parsing input files to the genelist format used by the RobustRankAggregation.
}
