% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetPP.R
\docType{package}
\name{TargetPP}
\alias{TargetPP}
\title{TargetPP: A package for integrating multiple CRISPR screens to prioritize
multi-phenotype targets}
\description{
The TargetPP package provides functions to integrate multiple CRISPR screens into
a single target-ranking and automatically generate summary and result plots.
}
\section{screen integration}{
 ...
}

\section{plotting}{
 ...
}

