% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_screens.R
\name{integrate_screens}
\alias{integrate_screens}
\title{CRISPR multi-screen integration}
\usage{
integrate_screens(
  datapath,
  infofile,
  inputformat = "CASPR",
  plot = TRUE,
  RRAexact = TRUE
)
}
\arguments{
\item{datapath}{a character string indicating the path to the folder
containing all CRISPR single-screen results.}

\item{infofile}{a character string indicating the path to the CRISPR
single-screen infofile.}

\item{inputformat}{a character string indicating the format of the input
data. One of "CASPR" (default), or "MAGeCK".}

\item{plot}{logical. Should summary plots be generated automatically?}

\item{RRAexact}{logical. When `TRUE` the RRA algorithm uses their "exact"
method to correct the p-value. Otherwise Bonferroni correction is applied.}
}
\value{
data.frame
}
\description{
\code{integrate_screens} integrates all CRISPR single-screen results in
\code{datapath} and returns a single target-ranking, based on significance
and effect size.
}
